% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.credible.intervals.R
\name{get.cred.intervals}
\alias{get.cred.intervals}
\title{Derives the credible intervals for marginal distributions of fixed-effect variables within an INLA model.}
\usage{
get.cred.intervals(Model, Quantiles = NULL)
}
\arguments{
\item{Model}{An INLA model of class "inla" that contains fixed effects.}

\item{Quantiles}{A numeric vector listing the credible intervals of interest. Values must be between 0 and 1. Default vector is c(0.5, 0.8, 0.9)}
}
\value{
Outputs a dataframe containing the fixed effects and the upper and lower bounds of the credibility intervals specified.
}
\description{
Derives the credible intervals for marginal distributions of fixed-effect variables within an INLA model.
}
\examples{
\donttest{
library(INLA)
Intercept <- 1
Slope <- 0.5
x <- seq(0, 10, length.out = 1000)
y <- Intercept + Slope * x + rnorm(1000, 0, 1)
df <- data.frame(x, y)
formula <- y ~ x
Model <- inla(formula, family = "normal", data = df)
get.cred.intervals(Model)
}
}
