% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Var.plot.R
\name{Var.plot}
\alias{Var.plot}
\title{Plots the credible intervals and mean of an INLA model's fixed effects.}
\usage{
Var.plot(
  Model,
  Plot_Intercept = FALSE,
  Plot_Title = NULL,
  Invert_Colors = FALSE,
  Quantiles = c(0.5, 0.8, 0.9),
  CI_width = c(2.5, 2, 1),
  Sig_threshold = NULL
)
}
\arguments{
\item{Model}{An INLA model of class "inla" that contains fixed effects.}

\item{Plot_Intercept}{Boolean (optional). Specifies whether to include the intercept in the resulting plot. Default is FALSE.}

\item{Plot_Title}{Character (optional). Specifies the title of the plot. Default is NULL.}

\item{Invert_Colors}{Boolean (optional). Specifies whether to invert the color palette when visualizing.}

\item{Quantiles}{A numeric vector (optional) listing the credible intervals of interest. Values must be between 0 and 1. Default vector is c(0.5, 0.8, 0.9). Must be the same length as the vector provided for CI_width.}

\item{CI_width}{A numeric vector (optional) listing the line size of the credible intervals. Default vector is c(2.5, 2, 1). Must be the same length as Quantiles.}

\item{Sig_threshold}{A numeric value (optional) specifying the credible interval threshold for coloring an effect size. Default is 0.9.}
}
\value{
Outputs a ggplot object of the fixed-effect variable effect sizes.
}
\description{
Plots the credible intervals and mean of an INLA model's fixed effects.
}
\examples{
\donttest{
library(INLA)
Intercept <- 1
Slope <- 0.5
x <- seq(0, 10, length.out = 1000)
y <- Intercept + Slope * x + rnorm(1000, 0, 1)
df <- data.frame(x, y)
formula <- y ~ x
Model <- inla(formula, family = "normal", data = df)
Var.plot(Model, Plot_Intercept = TRUE)
}
}
